requirements.txt
line-bot-sdk
flask
python-dotenv
openai